+++ examples/miniapps/commands-and-handlers/application/commands.py
@@ -1,0 +1,12 @@
+"""Commands module."""
+
+class Command:
+    ...
+
+
+class SaveRating(Command):
+    ...
+
+
+class DoSomethingElse(Command):
+    ...