+++ examples/miniapps/commands-and-handlers/application/repositories.py
@@ -1,0 +1,5 @@
+"""Repositories module."""
+
+
+class RatingRepository:
+    ...