+++ docs/main/changelog.rst
@@ -11,6 +11,8 @@ Develop
 -------
 - Add support of non-string keys for ``Dict`` provider.
 - Add simple ``FastAPI`` example.
+- Add ``Commands  and Handlers`` example from
+  issue `#327 <https://github.com/ets-labs/python-dependency-injector/issues/327>`_.
 - Add extra typing test for provided instance of ``DependenciesContainer`` provider.
 
 4.4.1